pyproject.toml
[project]
name = "bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "couchdb>=1.2",
    "elasticsearch==8.19.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
